@Component decorator . Declaring an @Component decorator that will be used in every component class in order. . Moving the rendering logic of a component to the @Component decorator
