+++ index.html
@@ -12,7 +12,7 @@
 			<article>
 				<header>
 					<h1>Building Angular from scratch</h1>
-					<div user-profile first-name="Mebarek" last-name="Zouakh" job="Developer">User Profile works!</div>
+					<user-profile first-name="Mebarek" last-name="Zouakh" job="Developer">User Profile works!</user-profile>
 				</header>
 				<input phone-number [will-have-spaces]="true" border-color="orange" type="text" name="phone-number" id="phone-number" placeholder="Numéro mobile" />
 				<input phone-number [will-have-spaces]="false" type="text" name="fix-number" id="fix-number" placeholder="Numéro fixe" />
